+++ R/AllGenerics.R
@@ -3,13 +3,6 @@ NULL
 
 setGeneric("backendBpparam", def = function(object, ...)
     standardGeneric("backendBpparam"))
-setGeneric("backendInitialize", def = function(object, ...)
-    standardGeneric("backendInitialize"),
-    valueClass = "MsBackend")
-#' @rdname hidden_aliases
-setGeneric("backendMerge", def = function(object, ...)
-    standardGeneric("backendMerge"),
-    valueClass = "MsBackend")
 #' @rdname hidden_aliases
 setGeneric("backendParallelFactor", def = function(object, ...)
     standardGeneric("backendParallelFactor"))
@@ -34,32 +27,9 @@ setGeneric("export", function(object, ...)
     standardGeneric("export"))
 setGeneric("filterFourierTransformArtefacts", function(object, ...)
     standardGeneric("filterFourierTransformArtefacts"))
-#' @rdname hidden_aliases
-setGeneric("filterMzRange", function(object, ...)
-    standardGeneric("filterMzRange"))
-#' @rdname hidden_aliases
-setGeneric("filterMzValues", function(object, ...)
-    standardGeneric("filterMzValues"))
-#' @rdname hidden_aliases
-setGeneric("filterPrecursorMzValues", function(object, ...)
-    standardGeneric("filterPrecursorMzValues"))
-#' @rdname hidden_aliases
-setGeneric("filterPrecursorMzRange", function(object, ...)
-    standardGeneric("filterPrecursorMzRange"))
-#' @rdname hidden_aliases
-setGeneric("isReadOnly", function(object, ...)
-    standardGeneric("isReadOnly"))
 #' @rdname neutralLoss
 setGeneric("neutralLoss", function(object, param, ...)
     standardGeneric("neutralLoss"))
-#' @rdname hidden_aliases
-setGeneric("peaksData", function(object, ...) standardGeneric("peaksData"))
-#' @rdname hidden_aliases
-setGeneric("peaksData<-", function(object, value)
-    standardGeneric("peaksData<-"))
-#' @rdname hidden_aliases
-setGeneric("peaksVariables", function(object, ...)
-    standardGeneric("peaksVariables"))
 #' @rdname hidden_aliases
 setGeneric("pickPeaks", function(object, ...)
     standardGeneric("pickPeaks"))
@@ -74,12 +44,7 @@ setGeneric("reset", function(object, ...)
 #' @rdname hidden_aliases
 setGeneric("selectSpectraVariables", function(object, ...)
     standardGeneric("selectSpectraVariables"))
-#' @rdname hidden_aliases
-setGeneric("setBackend", function(object, backend, ...)
-    standardGeneric("setBackend"))
 setGeneric("Spectra", function(object, ...) standardGeneric("Spectra"))
-#' @rdname hidden_aliases
-setGeneric("spectrapply", function(object, ...) standardGeneric("spectrapply"))
 
 #' @title Mapping between spectra variables and data file fields
 #'