+++ R/Spectra.R
@@ -1553,6 +1553,8 @@ setMethod("Spectra", "ANY", function(object, processingQueue = list(),
 
 #' @rdname Spectra
 #'
+#' @importMethodsFrom ProtGenerics setBackend
+#'
 #' @exportMethod setBackend
 setMethod(
     "setBackend", c("Spectra", "MsBackend"),
@@ -2234,6 +2236,8 @@ setMethod("filterMsLevel", "Spectra", function(object, msLevel. = integer()) {
 
 #' @rdname Spectra
 #'
+#' @importMethodsFrom ProtGenerics filterMzRange
+#'
 #' @export
 setMethod("filterMzRange", "Spectra",
           function(object, mz = numeric(), msLevel. = uniqueMsLevels(object),
@@ -2254,6 +2258,8 @@ setMethod("filterMzRange", "Spectra",
           })
 
 #' @rdname Spectra
+#'
+#' @importMethodsFrom ProtGenerics filterMzValues
 #'
 #' @export
 setMethod("filterMzValues", "Spectra",